Feature/tc 111

+++ src/main/java/codeenthusiast/TrainingCenterApp/exercise/enduranceexercise/EnduranceExerciseDetails.java
@@ -14,8 +14,8 @@ import javax.persistence.*;
 @AllArgsConstructor
 public class EnduranceExerciseDetails {
 
-    private int heartRate;
+    private short heartRate;
 
-    private int vO2maxPercentage;
+    private short vO2maxPercentage;
 
 }
